+++ example/android/app/src/main/kotlin/com/example/flutter_app_name_localization_example/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.example.flutter_app_name_localization_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity()